+++ security/specs/src/test/java/org/seedstack/seed/security/X509CertificateTokenTest.java
@@ -10,18 +10,20 @@ package org.seedstack.seed.security;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 import java.security.cert.X509Certificate;
+import javax.security.auth.x500.X500Principal;
 import org.junit.Test;
 
 public class X509CertificateTokenTest {
-
     @Test
     public void testToken() {
         X509Certificate certificate = mock(X509Certificate.class);
+        X500Principal x500Principal = new X500Principal("CN=Duke, OU=JavaSoft, O=Sun Microsystems, C=US");
+        when(certificate.getSubjectX500Principal()).thenReturn(x500Principal);
         X509CertificateToken token = new X509CertificateToken(new X509Certificate[]{certificate});
-        assertThat(token.getAuthenticatingCertificates()).containsExactly(certificate);
-        assertThat((X509Certificate[]) token.getCredentials()).containsExactly(certificate);
-        assertThat((X509Certificate[]) token.getPrincipal()).containsExactly(certificate);
+        assertThat(token.getCredentials()).isEqualTo(certificate);
+        assertThat(token.getPrincipal()).isEqualTo(x500Principal);
     }
 }