+++ web/undertow/src/test/java/org/seedstack/seed/undertow/UndertowIT.java
@@ -10,11 +10,11 @@ package org.seedstack.seed.undertow;
 
 import com.google.inject.Injector;
 import io.restassured.RestAssured;
+import io.restassured.config.SSLConfig;
 import io.restassured.response.Response;
 import org.assertj.core.api.Assertions;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 import org.seedstack.seed.Configuration;
 import org.seedstack.seed.core.Seed;
@@ -27,11 +27,6 @@ public class UndertowIT {
     private final SeedLauncher launcher = Seed.getLauncher();
     @Configuration("web.runtime.baseUrl")
     private String baseUrl;
-
-    @BeforeClass
-    public static void setUp() {
-        RestAssured.useRelaxedHTTPSValidation();
-    }
 
     @Before
     public void before() throws Exception {
@@ -59,7 +54,12 @@ public class UndertowIT {
     }
 
     private void checkServer() {
-        Response servletResponse = RestAssured.expect().statusCode(200).when().get(baseUrl + "hello");
+        Response servletResponse = RestAssured.given()
+                .config(RestAssured.config().sslConfig(SSLConfig.sslConfig().relaxedHTTPSValidation("SSL")))
+                .expect()
+                .statusCode(200)
+                .when()
+                .get(baseUrl + "hello");
         Assertions.assertThat(servletResponse.asString()).isEqualTo("Hello World!");
     }
 }